+++ drush.10.services.yml
@@ -1,7 +1,7 @@
 ---
 services:
   islandora_drush_utils.deleter:
-    class: \Drupal\islandora_drush_utils\Commands\Deleter
+    class: \Drupal\islandora_drush_utils\Drush\Commands\Deleter
     arguments:
       - '@entity_type.manager'
       - '@queue'
@@ -10,35 +10,35 @@ services:
     tags:
       - name: drush.command
   islandora_drush_utils.rederiver:
-    class: \Drupal\islandora_drush_utils\Commands\Rederive
+    class: \Drupal\islandora_drush_utils\Drush\Commands\Rederive
     arguments:
       - '@islandora.utils'
       - '@entity_type.manager'
     tags:
       - name: drush.command
   islandora_drush_utils.rederive_thumbnails:
-    class: \Drupal\islandora_drush_utils\Commands\GenerateThumbnails
+    class: \Drupal\islandora_drush_utils\Drush\Commands\GenerateThumbnails
     arguments:
       - '@entity_type.manager'
       - '@database'
     tags:
       - name: drush.command
   islandora_drush_utils.command.user_wrapper:
-    class: \Drupal\islandora_drush_utils\Commands\UserWrapperCommands
+    class: \Drupal\islandora_drush_utils\Drush\Commands\UserWrapperCommands
     arguments:
       - '@account_switcher'
       - '@entity_type.manager'
     tags:
       - name: drush.command
   islandora_drush_utils.command.user_wrapping_alterer:
-    class: \Drupal\islandora_drush_utils\Commands\UserWrappingAlterer
+    class: \Drupal\islandora_drush_utils\Drush\CommandInfoAlterers\UserWrappingAlterer
     tags:
       - name: drush.command_info_alterer
     arguments:
       - '@logger.islandora_drush_utils'
       - false
   islandora_drush_utils.null_child_weight_updater:
-    class: \Drupal\islandora_drush_utils\Commands\NullChildWeight
+    class: \Drupal\islandora_drush_utils\Drush\Commands\NullChildWeight
     arguments:
       - '@entity_type.manager'
       - '@database'
@@ -46,20 +46,20 @@ services:
     tags:
       - name: drush.command
   islandora_drush_utils.missing_derivatives:
-    class: \Drupal\islandora_drush_utils\Commands\MissingDerivatives
+    class: \Drupal\islandora_drush_utils\Drush\Commands\MissingDerivatives
     arguments:
       - '@entity_type.manager'
       - '@dgi_standard_derivative_examiner.examiner'
     tags:
       - name: drush.command
   islandora_drush_utils.derivatives_generator:
-    class: \Drupal\islandora_drush_utils\Commands\DerivativesGenerator
+    class: \Drupal\islandora_drush_utils\Drush\Commands\DerivativesGenerator
     arguments:
       - '@entity_type.manager'
     tags:
       - name: drush.command
   islandora_drush_utils.bulk_publish_unpublish:
-    class: \Drupal\islandora_drush_utils\Commands\PublishUnpublishCollections
+    class: \Drupal\islandora_drush_utils\Drush\Commands\PublishUnpublishCollections
     arguments:
       - '@entity_type.manager'
       - '@islandora.utils'
